Moved Testcases to the entities dir
