+++ tests/protocols/gmpv208/entities/notes/test_clone_note.py
@@ -16,27 +16,20 @@
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-from gvm.errors import GvmError
+from gvm.errors import RequiredArgument
 
 
-class GmpDeleteOverrideTestCase:
-    def test_delete(self):
-        self.gmp.delete_override('a1')
-
-        self.connection.send.has_been_called_with(
-            '<delete_override override_id="a1" ultimate="0"/>'
-        )
-
-    def test_delete_ultimate(self):
-        self.gmp.delete_override('a1', ultimate=True)
+class GmpCloneNoteTestMixin:
+    def test_clone(self):
+        self.gmp.clone_note('a1')
 
         self.connection.send.has_been_called_with(
-            '<delete_override override_id="a1" ultimate="1"/>'
+            '<create_note>' '<copy>a1</copy>' '</create_note>'
         )
 
     def test_missing_id(self):
-        with self.assertRaises(GvmError):
-            self.gmp.delete_override(None)
+        with self.assertRaises(RequiredArgument):
+            self.gmp.clone_note('')
 
-        with self.assertRaises(GvmError):
-            self.gmp.delete_override('')
+        with self.assertRaises(RequiredArgument):
+            self.gmp.clone_note(None)
