+++ tests/protocols/gmpv208/entities/notes/test_delete_note.py
@@ -16,20 +16,27 @@
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-from gvm.errors import RequiredArgument
+from gvm.errors import GvmError
 
 
-class GmpCloneNoteTestCase:
-    def test_clone(self):
-        self.gmp.clone_note('a1')
+class GmpDeleteNoteTestMixin:
+    def test_delete(self):
+        self.gmp.delete_note('a1')
 
         self.connection.send.has_been_called_with(
-            '<create_note>' '<copy>a1</copy>' '</create_note>'
+            '<delete_note note_id="a1" ultimate="0"/>'
+        )
+
+    def test_delete_ultimate(self):
+        self.gmp.delete_note('a1', ultimate=True)
+
+        self.connection.send.has_been_called_with(
+            '<delete_note note_id="a1" ultimate="1"/>'
         )
 
     def test_missing_id(self):
-        with self.assertRaises(RequiredArgument):
-            self.gmp.clone_note('')
+        with self.assertRaises(GvmError):
+            self.gmp.delete_note(None)
 
-        with self.assertRaises(RequiredArgument):
-            self.gmp.clone_note(None)
+        with self.assertRaises(GvmError):
+            self.gmp.delete_note('')
