+++ tests/protocols/gmpv208/entities/notes/test_get_note.py
@@ -1,0 +1,41 @@
+# -*- coding: utf-8 -*-
+# Copyright (C) 2018-2021 Greenbone Networks GmbH
+#
+# SPDX-License-Identifier: GPL-3.0-or-later
+#
+# This program is free software: you can redistribute it and/or modify
+# it under the terms of the GNU General Public License as published by
+# the Free Software Foundation, either version 3 of the License, or
+# (at your option) any later version.
+#
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License
+# along with this program.  If not, see <http://www.gnu.org/licenses/>.
+
+from gvm.errors import RequiredArgument
+
+
+class GmpGetNoteTestMixin:
+    def test_get_note(self):
+        self.gmp.get_note('n1')
+
+        self.connection.send.has_been_called_with(
+            '<get_notes note_id="n1" details="1"/>'
+        )
+
+        self.gmp.get_note(note_id='n1')
+
+        self.connection.send.has_been_called_with(
+            '<get_notes note_id="n1" details="1"/>'
+        )
+
+    def test_get_note_missing_note_id(self):
+        with self.assertRaises(RequiredArgument):
+            self.gmp.get_note(note_id=None)
+
+        with self.assertRaises(RequiredArgument):
+            self.gmp.get_note('')
